Struct rctree::Node

pub struct Node<T>(/* private fields */);

A reference to a node holding a value of type T. Nodes form a tree.

Internally, this uses reference counting for lifetime tracking and std::cell::RefCell for interior mutability.

Note: Cloning a Node only increments a reference count. It does not copy the data.

Implementations

impl<T> Node<T>

pub fn new(data: T) -> Node<T>

Creates a new node from its associated data.

pub fn downgrade(&self) -> WeakNode<T>

Returns a weak referece to a node.

pub fn parent(&self) -> Option<Node<T>>

Returns a parent node, unless this node is the root of the tree.

Panics

Panics if the node is currently mutably borrowed.

pub fn first_child(&self) -> Option<Node<T>>

Returns a first child of this node, unless it has no child.

Panics

Panics if the node is currently mutably borrowed.

pub fn last_child(&self) -> Option<Node<T>>

Returns a last child of this node, unless it has no child.

Panics

Panics if the node is currently mutably borrowed.

pub fn previous_sibling(&self) -> Option<Node<T>>

Returns the previous sibling of this node, unless it is a first child.

Panics

Panics if the node is currently mutably borrowed.

pub fn next_sibling(&self) -> Option<Node<T>>

Returns the next sibling of this node, unless it is a last child.

Panics

Panics if the node is currently mutably borrowed.

pub fn borrow(&self) -> Ref<'_, T>

Returns a shared reference to this nodeโs data

Panics

Panics if the node is currently mutably borrowed.

pub fn borrow_mut(&self) -> RefMut<'_, T>

Returns a unique/mutable reference to this nodeโs data

Panics

Panics if the node is currently borrowed.

pub fn ancestors(&self) -> Ancestors<T>

Returns an iterator of nodes to this node and its ancestors.

Includes the current node.

pub fn preceding_siblings(&self) -> PrecedingSiblings<T>

Returns an iterator of nodes to this node and the siblings before it.

Includes the current node.

pub fn following_siblings(&self) -> FollowingSiblings<T>

Returns an iterator of nodes to this node and the siblings after it.

Includes the current node.

pub fn children(&self) -> Children<T>

Returns an iterator of nodes to this nodeโs children.

Panics

Panics if the node is currently mutably borrowed.

pub fn has_children(&self) -> bool

Returns true if this node has children nodes.

Panics

Panics if the node is currently mutably borrowed.

pub fn descendants(&self) -> Descendants<T>

Returns an iterator of nodes to this node and its descendants, in tree order.

Includes the current node.

pub fn traverse(&self) -> Traverse<T>

Returns an iterator of nodes to this node and its descendants, in tree order.

pub fn detach(&self)

Detaches a node from its parent and siblings. Children are not affected.

Panics

Panics if the node or one of its adjoining nodes is currently borrowed.

pub fn append(&self, new_child: Node<T>)

Appends a new child to this node, after existing children.

Panics

Panics if the node, the new child, or one of their adjoining nodes is currently borrowed.

pub fn prepend(&self, new_child: Node<T>)

Prepends a new child to this node, before existing children.

Panics

Panics if the node, the new child, or one of their adjoining nodes is currently borrowed.

pub fn insert_after(&self, new_sibling: Node<T>)

Inserts a new sibling after this node.

Panics

Panics if the node, the new sibling, or one of their adjoining nodes is currently borrowed.

pub fn insert_before(&self, new_sibling: Node<T>)

Inserts a new sibling before this node.

Panics

Panics if the node, the new sibling, or one of their adjoining nodes is currently borrowed.

pub fn make_copy(&self) -> Node<T>
where T: Clone,

Returns a copy of a current node without children.

Panics

Panics if the node is currently mutably borrowed.

pub fn make_deep_copy(&self) -> Node<T>
where T: Clone,

Returns a copy of a current node with children.

Panics

Panics if any of the descendant nodes are currently mutably borrowed.